pyproject.toml
[project]
name = "english"
version = "0.1.0"
description = ""
authors = [
    {name = "hunhoon21",email = "hunhoon21@gmail.com"}
]
readme = "README.md"
requires-python = "^3.10"
dependencies = [
    "openai-whisper (>=20240930,<20240931)",
    "dotenv (>=0.9.9,<0.10.0)",
    "openai (>=1.70.0,<2.0.0)",
    "srt (>=3.5.3,<4.0.0)",
    "yt-dlp (>=2025.3.31,<2026.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
